* l7na, example: мелкие важные правки; * build: добавил boost_thread;

+++ src/l7na/drives.h
@@ -86,7 +86,7 @@ struct AxisStatus {
     AxisState   state;                  //!< Текущее состояние системы управления осью
     uint32_t    error_code;             //!< Код ошибки двигателя по CiA402
     uint16_t    statusword;             //!< Битовая маска текущего состояния привода (для отладки)
-    uint8_t     mode;                   //!< Текущий режим работы (для отладки)
+    uint16_t     mode;                   //!< Текущий режим работы (для отладки)
 };
 
 enum SystemState : int32_t {
